+++ CHANGELOG.markdown
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 * Changed the `@unit` built-in variable to be considered non-volatile and added proper handling of side effects by the `ubind` instruction. The change may allow some additional optimizations of unit handling code.
 * Changed linked variables to no longer be volatile while still protecting them in conditions. Explicitly declaring linked variables and arrays `volatile` is possible ([#327](https://github.com/cardillan/mindcode/issues/327)).
+* Changed the dynamic optimization process: when the current instruction count exceeds the limit, `neutral` and `speed` optimization goals are overriden to `size`. 
 * Temporary variables are renumbered at the end of compilation, so that their names (`*tmp#`) start with 0 and the sequence doesn't contain gaps.
 * Changed the program signature to include Mindcode version number.